requirements.txt
telethon>=1.36.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
requests>=2.28.0
